ToolBox-Ver 1.30.3.2.1

+++ controllers/usuarios.js
@@ -136,6 +136,11 @@ userRouter.put('/permisos/rol/:id', async (req, res) => {
             return res.status(404).json({ success: false, message: 'Usuario no encontrado' });
         }
 
+        // Impedir cambiar el rol de un usuario con rol 'boss'
+        if (user.rol === 'boss') {
+            return res.status(403).json({ success: false, message: 'No puedes cambiar el rol de un usuario con rol Boss' });
+        }
+
         user.rol = rol;
         await user.save();
 
@@ -151,10 +156,17 @@ userRouter.delete('/permisos/banear/:id', async (req, res) => {
     const userId = req.params.id;
 
     try {
-        const user = await User.findByIdAndDelete(userId);
+        const user = await User.findById(userId);
         if (!user) {
             return res.status(404).json({ success: false, message: 'Usuario no encontrado' });
         }
+
+        // Impedir banear a un usuario con rol 'boss'
+        if (user.rol === 'boss') {
+            return res.status(403).json({ success: false, message: 'No puedes banear a un usuario con rol Boss' });
+        }
+
+        await User.findByIdAndDelete(userId);
 
         res.json({ success: true, message: 'Usuario baneado correctamente' });
     } catch (error) {
